Revert "Front/fix/general issues (#306)"
This reverts commit 005a386da59677c15259f82e5f42e6c48d52b343.
