+++ front/client/components/AddCategoryDialog.vue
@@ -37,7 +37,6 @@
           <v-spacer></v-spacer>
           <v-btn
             flat
-            title="Cancel"
             color="primary"
             @click.native="close"
           >
@@ -45,7 +44,6 @@
           </v-btn>
           <v-btn
             flat
-            title="Submit"
             color="primary"
             class="add-category-submit-btn"
             :disabled="!isValid"
