+++ front/client/components/AddItemDialog.vue
@@ -31,7 +31,6 @@
         <v-btn
           flat
           color="primary"
-          title="Submit"
           class="add-cat-submit"
           :disabled="!isValid"
           @click.native="submit"
@@ -40,7 +39,6 @@
         </v-btn>
         <v-btn
           flat
-          title="Cancel"
           color="primary"
           @click.native="close"
         >
