+++ front/client/components/Categories.vue
@@ -75,7 +75,6 @@
           <v-btn
             class="ma-0 px-1"
             color="grey"
-            title="Add new category"
             flat
             @click="openAddCategoryDialog(groupName)"
           >
