+++ front/client/components/Category.vue
@@ -36,7 +36,6 @@
         flat
         class="ml-2"
         color="grey"
-        title="Add new item"
         @click="openAddItemDialog"
       >
         <v-icon size="14" class="mr-1" left>$vuetify.icons.plus</v-icon>
@@ -67,7 +66,7 @@ import AddItemDialog from 'client/components/AddItemDialog.vue'
     CategoryItem,
     CategoryDescription,
     CategoryInfo,
-    AddItemDialog
+    AddItemDialog,
   }
 })
 export default class Category extends Vue {
@@ -81,6 +80,18 @@ export default class Category extends Vue {
 
   get categoryUrl () {
     return this.category && `${_toKebabCase(this.category.title)}-${this.category.id}`
+  }
+
+  // TODO handle case when category was deleted. Go back in that case
+  async serverPrefetch () {
+    if (!this.categoryId) {
+      return
+    }
+    await this.$store.dispatch('category/loadCategory', this.categoryId)
+  }
+
+  beforeDestroy () {
+    this.$store.commit('category/setCategory', null)
   }
 
   openAddItemDialog () {
