+++ front/client/components/CategoryInfoEdit.vue
@@ -61,7 +61,6 @@
         <v-btn
           flat
           color="primary"
-          title="Submit"
           class="add-cat-submit"
           :disabled="!isValid"
           @click="updateCategoryInfo"
@@ -70,7 +69,6 @@
         </v-btn>
         <v-btn
           flat
-          title="Cancel"
           color="primary"
           @click.native="close"
         >
