+++ front/client/components/ConfirmDialog.vue
@@ -20,7 +20,6 @@
         <v-spacer />
         <v-btn
           flat
-          :title="cancelBtnText"
           color="primary"
           @click.native="cancel"
         >
@@ -28,7 +27,6 @@
         </v-btn>
         <v-btn
           flat
-          :title="confirmBtnText"
           color="primary"
           @click.native="confirm"
         >
