+++ front/client/components/MarkdownEditor.vue
@@ -20,7 +20,6 @@
       <v-toolbar-items>
         <v-btn
           small
-          title="Save"
           class="mr-2 text-transform-none"
           @click="save"
         >
@@ -28,7 +27,6 @@
         </v-btn>
         <v-btn
           small
-          title="Cancel"
           class="text-transform-none"
           @click="cancel"
         >
